List of R packages SDC book
This file contains the names of packages used by the book, "Statistical Data Cleaning with Applications in R". 
In the file, the package names are in quotes, and each name is followed by a comma (except last name in list). This facilitates installing the packages by simply pasting the content of the file into an object during an R session, and then using install.packages().